Better Config object structure
